feature: save and get in firestore database

+++ app/src/main/java/com/myapplication/domain/repository/MovieRepository.kt
@@ -25,7 +25,7 @@ import kotlinx.coroutines.launch
 interface MovieRepository {
     suspend fun addMovie(movie: MovieDetail)
     suspend fun getAllMovies(page: Int, query: String): Response<List<TopRatedResultItem>>
-    suspend fun getMovieDetails(movieId: Int): Response<MovieDetail>
+    suspend fun getMovieDetails(movieId: Long): Response<MovieDetail>
     suspend fun updateMovie(movie: TopRatedResult): Int
     suspend fun deleteMovie(movie: TopRatedResult): Int
 
@@ -91,7 +91,7 @@ class MovieDataSource(
         moviesDao.insertMovie(topRatedMovies)
     }
 
-    override suspend fun getMovieDetails(movieId: Int): Response<MovieDetail> {
+    override suspend fun getMovieDetails(movieId: Long): Response<MovieDetail> {
         return try {
             val movieDetail = service.getMovieDetails(movieId, Constants.API_KEY, DEFAULT_QUERY)
             CoroutineScope(Dispatchers.IO).launch {
